refactor: split custom data modules
