+++ vectorbt/vectorbt/data/__init__.py
@@ -4,7 +4,11 @@
 """Modules for working with data sources."""
 
 from vectorbt.data.base import symbol_dict, Data
-from vectorbt.data.custom import SyntheticData, GBMData, YFData, BinanceData, CCXTData, AlpacaData
+from vectorbt.data.synthetic import SyntheticData, GBMData
+from vectorbt.data.yfinance import YFData
+from vectorbt.data.binance import BinanceData
+from vectorbt.data.ccxt import CCXTData
+from vectorbt.data.alpaca import AlpacaData
 from vectorbt.data.updater import DataUpdater
 
 __all__ = [
